v3.7

+++ LiveHTS.Core/Interfaces/Model/IPersonContact.cs
@@ -4,7 +4,7 @@ namespace LiveHTS.Core.Interfaces.Model
 {
     public interface IPersonContact
     {
-        int? Phone { get; set; }
+        long? Phone { get; set; }
         bool Preferred { get; set; }
     }
 }
