+++ LiveHTS.Core/Model/Subject/PersonContact.cs
@@ -8,7 +8,7 @@ namespace LiveHTS.Core.Model.Subject
 {
     public class PersonContact : Entity<Guid>, IPersonContact
     {
-        public int? Phone { get; set; }
+        public long? Phone { get; set; }
         public bool Preferred { get; set; }
         [Indexed]
         public Guid PersonId { get; set; }
@@ -18,14 +18,14 @@ namespace LiveHTS.Core.Model.Subject
             Id = LiveGuid.NewGuid();
         }
 
-        private PersonContact(int? phone, bool preferred, Guid personId):this()
+        private PersonContact(long? phone, bool preferred, Guid personId):this()
         {
             Phone = phone;
             Preferred = preferred;
             PersonId = personId;
         }
 
-        public static PersonContact Create(int? phone, bool preferred, Guid personId)
+        public static PersonContact Create(long? phone, bool preferred, Guid personId)
         {
             return new PersonContact(phone, preferred, personId);
         }
